requirements.txt
fastapi
uvicorn
sentence-transformers
pydantic
numpy
python-dotenv
pyyaml
